# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/register": {
      "post": {
        "operationId": "UserController_registerUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User field updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Fields missing or validation error",
            "content": {
              "application/json": {
                "examples": {
                  "noName": {
                    "summary": "No name provided",
                    "value": {
                      "message": [
                        "Name should not be empty"
                      ],
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "noEmail": {
                    "summary": "No email provided",
                    "value": {
                      "message": [
                        "Email should not be empty"
                      ],
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registro de un nuevo usuario",
        "tags": [
          "User"
        ]
      }
    },
    "/users": {
      "put": {
        "operationId": "UserController_partialUpdate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserOptionalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User field updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "No fields provided to update",
            "content": {
              "application/json": {
                "example": {
                  "message": "No fields provided to update",
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid token",
                  "error": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "409": {
            "description": "No user can have the same email",
            "content": {
              "application/json": {
                "example": {
                  "message": "Email already exists",
                  "error": "Conflict",
                  "statusCode": 409
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "El usuario actualiza su información",
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All information of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "example": {
                  "message": "Invalid token",
                  "error": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "El usuario obtiene la información de su perfil",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "El usuario borra su perfil",
        "tags": [
          "User"
        ]
      }
    },
    "/users/report": {
      "post": {
        "operationId": "UserController_postReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "El usuario crea un nuevo reporte",
        "tags": [
          "User"
        ]
      }
    },
    "/users/reports": {
      "get": {
        "operationId": "UserController_getUserReports",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "El usuario obtiene todos los reportes que ha realizado",
        "tags": [
          "User"
        ]
      }
    },
    "/users/password": {
      "post": {
        "operationId": "UserController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Password changed successfully"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "examples": {
                  "invalidToken": {
                    "summary": "Invalid token",
                    "value": {
                      "message": "Invalid token",
                      "error": "Unauthorized",
                      "statusCode": 401
                    }
                  },
                  "invalidOldPassword": {
                    "summary": "Invalid old password",
                    "value": {
                      "message": "Invalid old password",
                      "error": "Unauthorized",
                      "statusCode": 401
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "El usuario cambia su contraseña",
        "tags": [
          "User"
        ]
      }
    },
    "/users/post-info": {
      "get": {
        "operationId": "UserController_getPostInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "pendiente": 1,
                  "aprobada": 3,
                  "rechazada": 0,
                  "total": 4,
                  "protegidas": 5
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "El usuario obtiene información sobre los reportes que ha realizado",
        "tags": [
          "User"
        ]
      }
    },
    "/users/settings-info": {
      "get": {
        "operationId": "UserController_getUserSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get user settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsUserDto"
                },
                "example": {
                  "is_reactions_enabled": 1,
                  "is_review_enabled": 1,
                  "is_reports_enabled": 1,
                  "is_anonymous_reports_enabled": 0
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "El usuario obtiene la configuración de su perfil",
        "tags": [
          "User"
        ]
      },
      "put": {
        "operationId": "UserController_updateSettingsInfo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Get user settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsUserDto"
                },
                "example": {
                  "is_reactions_enabled": 1,
                  "is_review_enabled": 1,
                  "is_reports_enabled": 1
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "error": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User with id 123 not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "El usuario cambia algún elemento de su configuración",
        "tags": [
          "User"
        ]
      }
    },
    "/reports": {
      "get": {
        "description": "Obtiene una lista de reportes con filtrado opcional por estado y paginación.",
        "operationId": "ReportsController_getReports",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "examples": {
              "pendiente": {
                "value": "1"
              },
              "aprobada": {
                "value": "2"
              },
              "rechazada": {
                "value": "3"
              }
            },
            "schema": {
              "enum": [
                "1",
                "2",
                "3"
              ],
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "examples": {
              "valid": {
                "value": "123"
              },
              "invalid": {
                "value": "abc"
              }
            },
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "examples": {
              "valid": {
                "value": 1
              },
              "invalid": {
                "value": -1
              }
            },
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reports retrieved successfully.",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "title": "Reporte de sitio fraudulento",
                    "image": "report-pictures/1d92a0a7cbb8f6a32b6ff1a98ecf2af4f13293be20e38e7db23ccfca9a412b8a.jpg",
                    "description": "Este sitio web solicita datos bancarios sin medidas de seguridad y redirige a páginas falsas.",
                    "created_at": "2025-09-27T23:50:39.000Z",
                    "updated_at": "2025-09-27T23:50:39.000Z",
                    "user_name": "Skibidi Toilet",
                    "created_by": 1,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://banco-seguro-falso.com",
                    "categories": [
                      1,
                      5
                    ]
                  },
                  {
                    "id": 2,
                    "title": "Página de phishing detectada",
                    "image": "report-pictures/34baf3de9a5ef873b432bd723d9e0d45a4c2b79f915a86d2a7b5419ccf78d66c.jpg",
                    "description": "El portal imita la interfaz de una empresa de envíos para robar credenciales de acceso.",
                    "created_at": "2025-09-27T23:51:32.000Z",
                    "updated_at": "2025-09-27T23:51:32.000Z",
                    "user_name": "Dr. Sahur",
                    "created_by": 2,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://envios-gratis-seguro.net",
                    "categories": [
                      2,
                      5
                    ]
                  },
                  {
                    "id": 3,
                    "title": "Sitio de ventas falsas",
                    "image": "report-pictures/57c0a3a48e923d8bb9e94b3a8ff743a58e9c4e71d4ccf0e6e2e3d513a7f49fdd.jpg",
                    "description": "La página ofrece productos electrónicos a precios demasiado bajos y nunca realiza las entregas.",
                    "created_at": "2025-09-27T23:51:51.000Z",
                    "updated_at": "2025-09-27T23:51:51.000Z",
                    "user_name": "Dr. Sahur",
                    "created_by": 3,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://ofertas-electronica-barata.org",
                    "categories": [
                      3,
                      5
                    ]
                  },
                  {
                    "id": 5,
                    "title": "Plataforma de inversión fraudulenta",
                    "image": "report-pictures/b9237d0f30ecb0a7dc2f84e6a1cc3af324ea742b93a56f5dd8f0cb11e629bf5e.jpg",
                    "description": "El sitio promete ganancias irreales en criptomonedas y exige depósitos sin garantía alguna.",
                    "created_at": "2025-09-27T23:52:29.000Z",
                    "updated_at": "2025-09-27T23:52:29.000Z",
                    "user_name": "Dr. Sahur",
                    "created_by": 2,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://crypto-ganancias-rapidas.info",
                    "categories": [
                      4,
                      5
                    ]
                  },
                  {
                    "id": 7,
                    "title": "Portal falso de soporte técnico",
                    "image": "report-pictures/ccf0a7bb2e9d9a4b8347ff291c4b731ddf28573c4a2a5b5d67e3bc111f0a89f3.jpg",
                    "description": "La web se hace pasar por un servicio de soporte oficial y pide pagos por reparaciones inexistentes.",
                    "created_at": "2025-09-27T23:52:50.000Z",
                    "updated_at": "2025-09-27T23:52:50.000Z",
                    "user_name": "Skibidi Toilet",
                    "created_by": 1,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://soporte-oficial-falso.com",
                    "categories": [
                      1,
                      4
                    ]
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "invalidStatus": {
                    "value": {
                      "message": "Status must be one of the following values: pendiente, aprobada, rechazada",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "invalidId": {
                    "value": {
                      "message": "ID must be a positive integer",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Report with specified ID not found.",
            "content": {
              "application/json": {
                "example": {
                  "message": "User with id 442 not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          }
        },
        "summary": "Obtiene los reportes",
        "tags": [
          "Reports"
        ]
      },
      "post": {
        "operationId": "ReportsController_createReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Create Report",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "reportId": 48,
                    "report_category": [
                      2,
                      3
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "examples": {
                  "missingFields": {
                    "summary": "Missing required fields",
                    "value": {
                      "statusCode": 400,
                      "message": "field should not be empty",
                      "error": "Bad Request"
                    }
                  },
                  "invalidCategory": {
                    "summary": "Invalid category",
                    "value": {
                      "statusCode": 400,
                      "message": "One or more categories are invalid",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Crea un nuevo reporte",
        "tags": [
          "Reports"
        ]
      }
    },
...
    "/admin/user/{id}": {
      "put": {
        "operationId": "AdminController_putUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "El id del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Los campos a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserOptionalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User field updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "No fields provided to update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "examples": {
                  "nofieldsToUpdate": {
                    "summary": "No fields provided to update",
                    "value": {
                      "message": "No fields provided to update",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "invalidEmail": {
                    "summary": "Email validation failed",
                    "value": {
                      "message": [
                        "email must be an email"
                      ],
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "passwordTooShort": {
                    "summary": "Password too short",
                    "value": {
                      "message": [
                        "password must be longer than or equal to 8 characters"
                      ],
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No user with that id found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "example": {
                  "message": "User with id 442 not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          },
          "409": {
            "description": "No user can have the same email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "example": {
                  "message": "Email already exists",
                  "error": "Conflict",
                  "statusCode": 409
                }
              }
            }
          }
        },
        "summary": "Modifica campos de un usuario por su id",
        "tags": [
          "Admin"
        ]
      },
      "get": {
        "operationId": "AdminController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "El id del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": 1,
                  "name": "Arturo Utrilla",
                  "email": "arturo@gmail.com",
                  "username": "Arturo_utrilla",
                  "password": "e28dce46d564d38b9b58f6c85ff4bb75ba3df1202991f5b6e1d5e5d585c1b8e7",
                  "salt": "9tuubdqpqei",
                  "created_at": "2025-09-25T12:14:44.000Z",
                  "updated_at": "2025-10-24T01:45:55.000Z",
                  "image_path": "profile-pictures/d84c32b6211caf458c18f49db1992bebbf5dbfccc581e47aad0b77eee3ea637a.jpg",
                  "role_id": 1
                }
              }
            }
          },
          "404": {
            "description": "No user with that id found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "example": {
                  "message": "User with id 442 not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          }
        },
        "summary": "Obtener un usuario por su id",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Endpoints de CRUD Usuarios",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "name@domain.com"
          },
          "name": {
            "type": "string",
            "example": "name"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "role_id": {
            "type": "string",
            "example": "1",
            "default": "1"
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
...
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "settings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSettingsDto"
            }
          }
        },
        "required": [
          "user",
          "settings"
        ]
      },
      "CreateUserOptionalDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "image_path": {
            "type": "string",
            "example": "/path/to/image.jpg"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "example": "55e1ebd3ebe4f1b46a5ccc9866d"
          },
          "salt": {
            "type": "string",
            "example": "432423"
          },
          "image_path": {
            "type": "string",
            "example": "/path/to/image.jpg"
          },
          "role_id": {
            "type": "number",
            "examples": [
              1,
              2
            ]
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "password",
          "salt",
          "role_id"
        ]
      },
...
      "ExceptionResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "type": "object"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "example": "oldPassword123"
          },
          "newPassword": {
            "type": "string",
            "example": "newPassword123"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
...
    }
  }
}